AP-5689 Improvement in Submenu (Correction based on recommendation)
